src/test/kotlin/be/inniger/advent/Day10Test.kt
package be.inniger.advent

import be.inniger.advent.util.readInputFile
import kotlin.test.Test
import kotlin.test.assertEquals

class Day10Test {

    private val input = readInputFile("10")

    @Test
    fun validateFirstSampleInputs() {
        val sampleLines = listOf(
            "[({(<(())[]>[[{[]{<()<>>",
            "[(()[<>])]({[<{<<[]>>(",
            "{([(<{}[<>[]}>{[]{[(<()>",
            "(((({<>}<{<{<>}{[]{[]{}",
            "[[<[([]))<([[{}[[()]]]",
            "[{[{({}]{}}([{[{{{}}([]",
            "{<[[]]>}<{[{[{[]{()[[[]",
            "[<(<(<(<{}))><([]([]()",
            "<{([([[(<>()){}]>(<<{{",
            "<{([{{}}[<[[[<>{}]]]>[]]"
        )
        assertEquals(26_397, Day10.solveFirst(sampleLines))
    }

    @Test
    fun validateFirstSolution() {
        assertEquals(271_245, Day10.solveFirst(input))
    }
}
